fix overflow

+++ 2.0/src/util.cu
@@ -101,7 +101,7 @@ int divide_ceil(int a, int b) {
 template <typename T>
 T* device_to_host(T* arr, int n) {
 	T* temp;
-	int tempBytes = sizeof(T) * n;
+	size_t tempBytes = sizeof(T) * n;
 	cudaMallocHost(&temp, tempBytes);
 	cudaMemcpy(temp, arr, tempBytes, cudaMemcpyDeviceToHost);
 	return temp;
@@ -110,7 +110,7 @@ T* device_to_host(T* arr, int n) {
 template <typename T>
 T* host_to_device(T* arr, int n) {
 	T* temp;
-	int tempBytes = sizeof(T) * n;
+	size_t tempBytes = sizeof(T) * n;
 	cudaMalloc(&temp, tempBytes);
 	cudaMemcpy(temp, arr, tempBytes, cudaMemcpyHostToDevice);
 	return temp;
